+++ src/main/java/techcourse/myblog/annotation/LoginUser.java
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface UserFromSession {
+public @interface LoginUser {
 }
