+++ src/main/java/techcourse/myblog/controller/CommentController.java
@@ -7,14 +7,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import techcourse.myblog.annotation.UserFromSession;
+import techcourse.myblog.annotation.LoginUser;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.Comment;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.NotFoundArticleException;
 import techcourse.myblog.exception.NotFoundCommentException;
-import techcourse.myblog.repository.ArticleRepository;
-import techcourse.myblog.repository.CommentRepository;
+import techcourse.myblog.domain.repository.ArticleRepository;
+import techcourse.myblog.domain.repository.CommentRepository;
 import techcourse.myblog.service.dto.CommentDTO;
 import techcourse.myblog.service.dto.CommentResponseDto;
 import techcourse.myblog.service.dto.ResponseMessage;
@@ -24,9 +24,9 @@ import javax.servlet.http.HttpServletRequest;
 @RestController
 @RequestMapping("/articles/{articleId}/comments")
 public class CommentController {
+    private static final Logger log = LoggerFactory.getLogger(CommentController.class);
     private final CommentRepository commentRepository;
     private final ArticleRepository articleRepository;
-    private static final Logger log = LoggerFactory.getLogger(CommentController.class);
 
     public CommentController(CommentRepository commentRepository, ArticleRepository articleRepository) {
         this.commentRepository = commentRepository;
@@ -36,7 +36,7 @@ public class CommentController {
     @PostMapping
     public CommentResponseDto create(@RequestBody CommentDTO commentDTO,
                                      @PathVariable long articleId,
-                                     @UserFromSession User user) {
+                                     @LoginUser User user) {
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(NotFoundArticleException::new);
         Comment comment = commentDTO.toDomain(article, user);
@@ -49,7 +49,7 @@ public class CommentController {
     @DeleteMapping("/{commentId}")
     public ResponseEntity<ResponseMessage> delete(@PathVariable long articleId,
                                                   @PathVariable long commentId,
-                                                  @UserFromSession User user) {
+                                                  @LoginUser User user) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(NotFoundCommentException::new);
         comment.validate(user);
@@ -57,7 +57,7 @@ public class CommentController {
                 .orElseThrow(NotFoundArticleException::new);
         article.remove(comment);
         commentRepository.delete(comment);
-        ResponseMessage responseMessage = new ResponseMessage("success", "", "", "");
+        ResponseMessage responseMessage = new ResponseMessage("success", "comment delete success", "", "");
         return new ResponseEntity<>(responseMessage, HttpStatus.OK);
     }
 
@@ -66,23 +66,21 @@ public class CommentController {
     public ResponseEntity<ResponseMessage> update(@PathVariable long articleId,
                                                   @PathVariable long commentId,
                                                   @RequestBody CommentDTO commentDTO,
-                                                  @UserFromSession User user) {
+                                                  @LoginUser User user) {
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(NotFoundCommentException::new);
         comment.validate(user);
 
         comment.update(commentDTO.toDomain(comment.getArticle(), comment.getAuthor()));
-        ResponseMessage responseMessage = new ResponseMessage("success", "", "", "");
-
+        ResponseMessage responseMessage = new ResponseMessage("success", "comment update success", "", "");
         return new ResponseEntity<>(responseMessage, HttpStatus.OK);
     }
 
     @ExceptionHandler(RuntimeException.class)
     public ResponseEntity<ResponseMessage> exceptionHandler(RuntimeException exception, RedirectAttributes redirectAttributes, HttpServletRequest request) {
         redirectAttributes.addFlashAttribute("commentError", exception.getMessage());
-        ResponseMessage responseMessage = new ResponseMessage("", "",
-                exception.getMessage(), "");
+        ResponseMessage responseMessage = new ResponseMessage("", "", exception.getMessage(), "");
         log.error("error : {}", exception.getMessage());
         return new ResponseEntity<>(responseMessage, HttpStatus.BAD_REQUEST);
     }
