+++ src/main/java/techcourse/myblog/controller/argumentresolver/UserArgumentResolver.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import techcourse.myblog.annotation.UserFromSession;
+import techcourse.myblog.annotation.LoginUser;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.UnauthenticatedUserException;
 
@@ -14,18 +14,16 @@ import javax.servlet.http.HttpServletRequest;
 public class UserArgumentResolver implements HandlerMethodArgumentResolver {
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(UserFromSession.class);
+        return parameter.hasParameterAnnotation(LoginUser.class);
     }
 
     @Override
     public User resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         User user = (User) request.getSession().getAttribute("user");
-
         if (user == null) {
             throw new UnauthenticatedUserException();
         }
-
         return user;
     }
 }
