+++ src/main/java/techcourse/myblog/service/LoginService.java
@@ -5,7 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.LoginFailException;
 import techcourse.myblog.exception.NotFoundUserException;
-import techcourse.myblog.repository.UserRepository;
+import techcourse.myblog.domain.repository.UserRepository;
 import techcourse.myblog.service.dto.UserDTO;
 
 @Service
