+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.EmailRepetitionException;
-import techcourse.myblog.repository.UserRepository;
+import techcourse.myblog.domain.repository.UserRepository;
 import techcourse.myblog.service.dto.UserDTO;
 
 import java.util.List;
